Split source and tests into two files to match data types
